+++ apps/docs/src/app/core/component-docs/object-marker/examples/object-marker-clickable-example.component.ts
@@ -1,0 +1,12 @@
+import { Component } from '@angular/core';
+
+@Component({
+    selector: 'fd-object-marker-clickable-example',
+    templateUrl: './object-marker-clickable-example.component.html'
+})
+export class ObjectMarkerClickableExampleComponent {
+    glyph = 'private';
+    changeGlyph(): void {
+        this.glyph = 'add-favorite';
+    }
+}
