+++ apps/docs/src/app/core/component-docs/object-marker/examples/object-marker-example.component.ts
@@ -17,14 +17,3 @@ export class ObjectMarkerIconAndTextExampleComponent {}
     templateUrl: './object-marker-text-example.component.html'
 })
 export class ObjectMarkerTextExampleComponent {}
-
-@Component({
-    selector: 'fd-object-marker-clickable-example',
-    templateUrl: './object-marker-clickable-example.component.html'
-})
-export class ObjectMarkerClickableExampleComponent {
-    glyph = 'private';
-    changeGlyph(): void {
-        this.glyph = 'add-favorite';
-    }
-}
