+++ apps/docs/src/app/core/component-docs/object-marker/object-marker-docs.component.html
@@ -14,7 +14,7 @@
     Object marker Text and clickable Text
 </fd-docs-section-title>
 <description>
-    Standard Object marker Text and clickable use <code>clickable</code> boolean property to add hoverig effects.
+    Standard Object marker Text and clickable use <code>clickable</code> boolean property to add hovering effects.
 </description>
 <component-example>
     <fd-object-marker-text-example></fd-object-marker-text-example>
@@ -33,10 +33,10 @@
 <separator></separator>
 
 <fd-docs-section-title id="clickable" componentName="objectmarker">
-    Object marker Icon and text clickable.
+    Object marker icon and text clickable.
 </fd-docs-section-title>
 <description>
-    Standard Object marker with icon and text with clickable use <code>clickable</code> boolean property to add hoverig
+    Standard Object marker with icon and text with clickable use <code>clickable</code> boolean property to add hovering
     effects.
 </description>
 <component-example>
