+++ libs/core/src/lib/object-marker/object-marker.component.scss
@@ -1,2 +1,1 @@
-@import '~fundamental-styles/dist/icon';
 @import '~fundamental-styles/dist/object-marker';
